pillar container: set proper root overlayf
We previously had no real root for containers, and then mounted a target overlayfs from xen-root to / in the container.
Newer versions of runc properly disallow this. Instead, we create an explicit overlayfs with xen-root as the lower,
and then make the merged filesystem be the root of the container.
